Move some tests to transmission specific class
