+++ sdk/appcenter-analytics/src/test/java/com/microsoft/appcenter/analytics/AnalyticsTransmissionTargetTest.java
@@ -11,13 +11,20 @@ import org.junit.Test;
 import org.mockito.ArgumentMatcher;
 import org.mockito.Mock;
 
+import java.util.HashMap;
+
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyString;
 import static org.mockito.Matchers.argThat;
+import static org.mockito.Matchers.isA;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.verify;
 
 public class AnalyticsTransmissionTargetTest extends AbstractAnalyticsTest {
@@ -34,6 +41,114 @@ public class AnalyticsTransmissionTargetTest extends AbstractAnalyticsTest {
         analytics.onStarting(mAppCenterHandler);
         analytics.onStarted(mock(Context.class), mChannel, null, null, false);
     }
+
+    @Test
+    public void testGetTransmissionTarget() {
+        assertNull(Analytics.getTransmissionTarget(""));
+        assertNotNull(Analytics.getTransmissionTarget("token"));
+    }
+
+    @Test
+    public void trackEventWithTransmissionTargetFromApp() {
+        testTrackEventWithTransmissionTarget("defaultToken", true);
+    }
+
+    @Test
+    public void trackEventWithTransmissionTargetFromLibrary() {
+        testTrackEventWithTransmissionTarget(null, false);
+    }
+
+    private void testTrackEventWithTransmissionTarget(final String defaultToken, boolean startFromApp) {
+        AnalyticsTransmissionTarget target = Analytics.getTransmissionTarget("token");
+        assertNotNull(target);
+
+        /* Getting a reference to the same target a second time actually returns the same. */
+        assertSame(target, Analytics.getTransmissionTarget("token"));
+
+        /* Track event with default transmission target. */
+        Analytics.trackEvent("name");
+        if (startFromApp) {
+            verify(mChannel).enqueue(argThat(new ArgumentMatcher<Log>() {
+
+                @Override
+                public boolean matches(Object item) {
+                    if (item instanceof EventLog) {
+                        EventLog eventLog = (EventLog) item;
+                        boolean nameAndPropertiesMatch = eventLog.getName().equals("name") && eventLog.getProperties() == null;
+                        boolean tokenMatch;
+                        if (defaultToken != null) {
+                            tokenMatch = eventLog.getTransmissionTargetTokens().size() == 1 && eventLog.getTransmissionTargetTokens().contains(defaultToken);
+                        } else {
+                            tokenMatch = eventLog.getTransmissionTargetTokens().isEmpty();
+                        }
+                        return nameAndPropertiesMatch && tokenMatch;
+                    }
+                    return false;
+                }
+            }), anyString());
+        } else {
+            verify(mChannel, never()).enqueue(isA(EventLog.class), anyString());
+        }
+        reset(mChannel);
+
+        /* Track event via transmission target method. */
+        target.trackEvent("name");
+        verify(mChannel).enqueue(argThat(new ArgumentMatcher<Log>() {
+
+            @Override
+            public boolean matches(Object item) {
+                if (item instanceof EventLog) {
+                    EventLog eventLog = (EventLog) item;
+                    boolean nameAndPropertiesMatch = eventLog.getName().equals("name") && eventLog.getProperties() == null;
+                    boolean tokenMatch = eventLog.getTransmissionTargetTokens().size() == 1 && eventLog.getTransmissionTargetTokens().contains("token");
+                    return nameAndPropertiesMatch && tokenMatch;
+                }
+                return false;
+            }
+        }), anyString());
+        reset(mChannel);
+
+        /* Track event via another transmission target method with properties. */
+        Analytics.getTransmissionTarget("token2").trackEvent("name", new HashMap<String, String>() {{
+            put("valid", "valid");
+        }});
+        verify(mChannel).enqueue(argThat(new ArgumentMatcher<Log>() {
+
+            @Override
+            public boolean matches(Object item) {
+                if (item instanceof EventLog) {
+                    EventLog eventLog = (EventLog) item;
+                    boolean nameAndPropertiesMatch = eventLog.getName().equals("name") && eventLog.getProperties().size() == 1 && "valid".equals(eventLog.getProperties().get("valid"));
+                    boolean tokenMatch = eventLog.getTransmissionTargetTokens().size() == 1 && eventLog.getTransmissionTargetTokens().contains("token2");
+                    return nameAndPropertiesMatch && tokenMatch;
+                }
+                return false;
+            }
+        }), anyString());
+        reset(mChannel);
+
+        /* Create a child transmission target and track event. */
+        AnalyticsTransmissionTarget childTarget = target.getTransmissionTarget("token3");
+        childTarget.trackEvent("name");
+        verify(mChannel).enqueue(argThat(new ArgumentMatcher<Log>() {
+
+            @Override
+            public boolean matches(Object item) {
+                if (item instanceof EventLog) {
+                    EventLog eventLog = (EventLog) item;
+                    boolean nameAndPropertiesMatch = eventLog.getName().equals("name") && eventLog.getProperties() == null;
+                    boolean tokenMatch = eventLog.getTransmissionTargetTokens().size() == 1 && eventLog.getTransmissionTargetTokens().contains("token3");
+                    return nameAndPropertiesMatch && tokenMatch;
+                }
+                return false;
+            }
+        }), anyString());
+        reset(mChannel);
+
+        /* Another child transmission target with the same token should be the same instance. */
+        assertSame(childTarget, target.getTransmissionTarget("token3"));
+    }
+
 
     @Test
     public void setEnabled() {
